requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-SocketIO==5.3.6
eventlet==0.36.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
Pillow==11.2.1
requests==2.31.0
pytesseract==0.3.13
python-dotenv==1.0.1
openpyxl==3.1.2
opencv-python-headless==4.10.0.84
